feat(ui): move ReceiveCredential subfolder
