+++ src/ui/pages/NotificationDetails/components/ReceiveCredential/ReceiveCredential.scss
