+++ src/ui/pages/NotificationDetails/components/ReceiveCredential/index.ts
@@ -1,0 +1,2 @@
+export * from "./ReceiveCredential";
+export * from "./ReceiveCrendential.test";
